Disable the -m prevention
Why you made the change:

I disabled this so `git rebase --continue` would still work properly.

How the change addresses the need:

Sadly, it seems that `git rebase --continue` is either using the `-m` or
`-F` option and therefore being represented as a "message" sourced
commit, which with the associated code would reject the commit.
Commenting out the code that rejects this allows `git rebase --continue`
to work properly.