Remove missing branch hard failure
Why you made the change:

To allow `git rebase --continue` to continue to work properly.

How the change addresses the need:

It seems that when in the middle of a rebase and you are headless
momentarily, and then do a `git rebase --continue` it was trying to grab
the branch name and hard failing. Therefore, enable to allow `git rebase
--continue` to continue working I removed this hard failure check and
allow it to continue.