Improve git commit message template
My intent is to improve the git commit message template in a number of
ways. First, correct spelling of INTENTION so that people find it easier
to read. Second, combine INTENTION and REASON sections into a singular
section as I feel it will be more natural to write about them together.
Third, trimmed the description headings down so that it will feel
lighter and in turn reduce the friction. Fourth, add section headings
for "Deployment or Rollout Considerations" and "Automated Tests
considerations" so that people would be reminded to think of those
things as well as provide necessary context around them.